% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{add}
\alias{add}
\title{A Cat Function}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function allows you to express your love of cats.
}
\examples{
Add together two numbers.

}
\keyword{cats}
