% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_areathreshold.R
\name{hsa_areathreshold}
\alias{hsa_areathreshold}
\title{Function for applying minimum area threshold}
\usage{
hsa_areathreshold(fdf, area_thr)
}
\arguments{
\item{fdf}{Flows data frame for water year in format of 'utils_hsaflws' function}

\item{area_thr}{Minimum area threshold}
}
\value{
Flows data frame with areaminreq field filled in.
}
\description{
This function accepts the flows data frame and calculates whether the area
threshold is exceeded at some point during a flood event.
}
\details{
If the area requirement is met for at least one day during a given
flood event, all days are given a value of one, otherwise the days are set to zero.
}
