% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_extent.R
\name{hsa_extent}
\alias{hsa_extent}
\title{Hydrospatial analysis of inundation extent with depth and velocity criteria}
\usage{
hsa_extent(rs_d, rs_v, fdf, dmin, dmax, vmax, wy, cres, aconv, fp_a, ncor,
  outdir)
}
\arguments{
\item{rs_d}{Raster stack or brick of depth}

\item{rs_v}{Raster stack or brick of velocity}

\item{fdf}{Flows data frame for water year in format of 'utils_hsaflws'
function}

\item{dmin}{Minimum depth threshold for raster calculation}

\item{dmax}{Maximum depth threshold for raster calculation}

\item{vmax}{Maximum velocity threshold for raster calculation}

\item{wy}{Water year to add to filenames}

\item{cres}{Resolution of cell (in units squared)}

\item{aconv}{Conversion factor for calculating area}

\item{fp_a}{Area of study site}

\item{ncor}{Number of cores for parallel processing}

\item{outdir}{Directory for writing rasters to file}
}
\value{
Flows data frame with inundated area metrics filled in. Writes the
  three sets of rasters to file in the outdir.
}
\description{
Accepts depth and velocity rasters (from \code{predrast_interp}) and
associated flows data frame for a given water year and calculates three sets
of binary rasters, indicating whether cells are inundated, whether they meet
the depth thresholds, and whether they meet the velocity threshold. The area
for each raster is determined and added to the flows data frame.
}
\details{
Rasters are written to file in a 'rsi' directory within 'outdir'.
}
