% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_freq.R
\name{hsa_freq}
\alias{hsa_freq}
\title{Hydrospatial analysis of inundation frequency}
\usage{
hsa_freq(rs_i0, fdf, wy, ncor, outdir)
}
\arguments{
\item{rs_i0}{Raster stack or brick with inundated cells = 1}

\item{fdf}{Flows data frame for water year in format of 'utils_hsaflws'
function}

\item{wy}{Water year to add to filenames}

\item{ncor}{Number of cores for parallel processing}

\item{outdir}{Directory for writing rasters to file}
}
\value{
Writes rasters with groupings of inundation to file in the outdir.
}
\description{
Accepts a raster stack/brick representing inundated area (from
\code{hsa_extent}) and assigns unique values per cell across the stack/brick
for each group of consecutive days of inundation.
}
\details{
Input rasters are in the 'rsi' format of output from
  \code{hsa_extent}, where cells = 1 are inundated (meeting depth and
  velocity threshold criteria) and non-inundated cells = 0. Rasters are
  written to file in a 'rsnoinun' directory within 'outdir'.
}
