% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_timing.R
\name{hsa_timing}
\alias{hsa_timing}
\title{Function for hydrospatial application of seasonal timing window}
\usage{
hsa_timing(rs_dayinunwgt, mo_start, mo_end, fdf, outdir)
}
\arguments{
\item{rs_dayinunwgt}{Raster stack or brick of inundation day as from 'hsa_duration'}

\item{mo_start}{Threshold start month}

\item{mo_end}{Threshold number of days for long duration inundation}

\item{fdf}{Flows data frame for water year in format of 'utils_hsaflws' function}

\item{outdir}{Directory for writing rasters to file}
}
\value{
Flood events data frame for water year with duration metrics filled in.
Writes the duration rasters to file in the outdir.
}
\description{
This function accepts
}
\details{
Input rasters are in the  Rasters are written to
file in a 'rsdur0' directory within 'outdir'.
}
