% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-hydrospatial.R
\docType{package}
\name{hydrospatial}
\alias{hydrospatial}
\alias{hydrospatial-package}
\title{hydrospatial: A package for characterizing floodplain inundation and habitat
in space and time}
\description{
An R-package for spatial and temporal analysis of floodplain inundation
characteristics and application of habitat suitability criteria.
It takes as input raster data of depth and velocity (e.g., from 2D hydrodynamic
modeling) at known flows and a corresponding daily flow time series.
Daily rasters of estimated depth and velocity are generated from
inputs using spatially resolved piece-wise linear interpolation.
Hydrospatial analysis functions calculate a range of physical metrics using
the raster time series. Habitat suitability criteria applied to the
computed physical metrics estimate spatially resolved daily cell suitability,
allowing for computation of daily weighted usable area. Statistical and
graphical summary of these metrics in space and time.
}
\details{
The hydrospatial package provides four categories of functions:
predictrast, hsa, utility, and vis
}
\section{predictrast functions}{

The predictrast functions
}

