% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predrast_interp.R
\name{predrast_interp}
\alias{predrast_interp}
\title{Create interpolated rasters based on a flow time series and rasters at known
flows}
\usage{
predrast_interp(rs_mod, r_ithr, flws_mod, flws_pred, ponding = FALSE, sc,
  vbl, preddir)
}
\arguments{
\item{rs_mod}{Raster stack or brick with each layer corresponding to the flws
vector}

\item{r_ithr}{Raster layer of flow inundation thresholds of cells, as from
the 'predrast_thresholds' function. Used if 'ponding' is TRUE.}

\item{flws_mod}{Data frame of flows corresponding to the raster stack or
brick, containing a 'flw' numeric column and a 'limb' attributing flows as
either rising ('r') or falling ('f')}

\item{flws_pred}{Data frame of flows for which to create interpolated
rasters, composed of a date ('dt') column, flows ('flw') column, a limb
('limb') column assigning the flow as either on the rising ('r') or falling
('f') limb of the hydrograph, and a high flow ('hflw') column representing
the recent peak flow to which ponded or disconnected areas correspond (such
as the highest flow within the last seven days).}

\item{ponding}{Set ponding to TRUE or FALSE. Default value is FALSE.}

\item{sc}{Character string to attach to the output raster filenames (such as
the name of the model scenario)}

\item{vbl}{Character string to attach to the output raster filenames (e.g.,
'Depth')}

\item{preddir}{Directory to which the rasters will be written}
}
\value{
Rasters written to file
}
\description{
Creates and writes to file a raster corresponding to each record in an input
flow time series, applying spatially-resolved piece-wise linear interpolation
using rasters at known flows.
}
\details{
This function accepts a stack of rasters (typically of inundation
  depth or velocity, as from 2D hydrodynamic modeling) representing
  conditions at known flows, which are used to create an interpolated raster
  for each flow in the input flow time series. The input flow time series is
  typically daily flow time series.

  If the ponding option is set to true, the function determines whether the
  flows in the input flow time series are on the rising or falling limb of
  the hydrograph and uses the corresponding rasters from the input raster
  stack. It also uses the flow threshold raster created by the
  'predrast_thresholds' function to determine whether cells should be wet or
  dry on the falling limb of a hydrograph based on the high flow column.
}
