% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areaday.R
\name{utils_areaday}
\alias{utils_areaday}
\title{Function to summarize area-day from daily results}
\usage{
utils_areaday(d, sumby)
}
\arguments{
\item{d}{Data frame in format used for 'hsa' functions}

\item{sumby}{Character field name to summarize on}
}
\value{
List of annual 'timeseries' and 'summary' of that time series
}
\description{
This function takes the daily hsa results table from hydrospatial analysis
and summarizes annual on a given field, returning a list of an annual
'timeseries' and a 'summary' of that time series
}
