% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_floodid.R
\name{utils_floodid}
\alias{utils_floodid}
\title{Function to identify flood days and events}
\usage{
utils_floodid(d, Q, outdir)
}
\arguments{
\item{d}{Data frame of daily flow time series}

\item{Q}{Vector of flow thresholds to use to id flood days}

\item{outdir}{Directory for writing flows to file}
}
\value{
List of 1) floods, 2) filenames, 3) daily flows with flood id.
Writes to file daily and flood event csvs.
}
\description{
This function accepts a daily flow time series and a vector of flood
thresholds for identifying flood days.
}
\details{
A daily flow time series data frame is required (as from
utils_flowformat function), including the following columns: dt (Date),
wyr (water year), wyrd (water year day), flw (flow), cflow (cumulative
flow to date). Using a single value or vector of flood thresholds, flood
days are identified and flood event metrics are determined.
}
