% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flowformat.R
\name{utils_flowformat}
\alias{utils_flowformat}
\title{Function to format daily flow time series}
\usage{
utils_flowformat(d)
}
\arguments{
\item{d}{A daily flow time series data frame including a date column
('dt') in mdy format and flow column ('flw'). There should be no missing
days.}
}
\value{
Formatted daily flows data frame
}
\description{
This function accepts a daily flow time series and returns a formatted
daily flow time series for use in other hydrospatial functions
}
