% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flowstopredict.R
\name{utils_flowstopredict}
\alias{utils_flowstopredict}
\title{Format flows for raster prediction}
\usage{
utils_flowstopredict(d, d_flds, thr, postfld = 7, outdir)
}
\arguments{
\item{d}{A daily flow time series data frame}

\item{d_flds}{Flood days data frame}

\item{thr}{Integer flow threshold to identify flood days}

\item{postfld}{Days post-flood to retain in dataset. Default is 7.}

\item{outdir}{Directory for writing flows to file}
}
\value{
Flows data frame as input for raster prediction, written to file as
  "flows_topred_full.csv"
}
\description{
Identifies flow days for hydrospatial analysis, to be used as input for
raster prediction
}
\details{
It takes a formatted flow time series and flood day flow time series
  from hydrospatial 'utils_flowformat' and 'utils_floodid' functions,
  respectively.
}
