% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hsaflws.R
\name{utils_hsaflws}
\alias{utils_hsaflws}
\title{Prepare daily flows table for hydrospatial analysis}
\usage{
utils_hsaflws(d, sc)
}
\arguments{
\item{d}{Data frame in format of 'utils_flowstopredict' function}

\item{sc}{Scenario name}
}
\value{
Flows data frame as input for hydrospatial analysis
}
\description{
Prepares the flows table that will be used in hydrospatial analysis, taking
as input the flows table from the 'utils_flowstopredict' function.
}
