% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hsaflwsevts.R
\name{utils_hsaflwsevts}
\alias{utils_hsaflwsevts}
\title{Prepare flood events table for hydrospatial analysis}
\usage{
utils_hsaflwsevts(d, sc)
}
\arguments{
\item{d}{Data frame in format of 'utils_hsaflws' function}

\item{sc}{Scenario name}
}
\value{
Flood events data frame as input for hydrospatial analysis
}
\description{
Prepares the flood events table that will be used in hydrospatial analysis,
taking as input the flows table from the 'utils_hsaflws' function.
}
