% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rasterreord.R
\name{utils_rasterreord}
\alias{utils_rasterreord}
\title{Function to reorder raster files}
\usage{
utils_rasterreord(fls)
}
\arguments{
\item{fls}{Vector of filenames with numbers to order after the last "_"}
}
\value{
Reordered filenames
}
\description{
This function prepares the flows table that will be used in
hydrospatial analysis, taking as input the flows table from the
'utils_flowstopredict' function.
}
