% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rsltsexprob.R
\name{utils_rsltsexprob}
\alias{utils_rsltsexprob}
\title{Set exceedance probability table from annual time series}
\usage{
utils_rsltsexprob(d, tmsrs, sc)
}
\arguments{
\item{d}{Data frame in format used for 'hsa' functions}

\item{tmsrs}{Data frame of annual time series from 'utils_areaday' function}

\item{sc}{Scenario name to add as a column}
}
\value{
Flows data frame as input for hydrospatial analysis
}
\description{
Prepares the exceedance probability table from time series results table for
use in visualizing.
}
