% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rsltsformat_daily.R
\name{utils_rsltsformat_daily}
\alias{utils_rsltsformat_daily}
\title{Format daily results into continuous time series}
\usage{
utils_rsltsformat_daily(d, daily, sc)
}
\arguments{
\item{d}{Data frame in format used for 'hsa' functions}

\item{daily}{Data frame of continous daily flow record}

\item{sc}{Scenario name to add as a column}
}
\value{
Flows data frame as input for hydrospatial analysis
}
\description{
Prepares the daily hsa results table from hydrospatial analysis and
transforms it into a continuous time series for use in visualizing.
}
