% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_connectivity.R
\name{hsa_connectivity}
\alias{hsa_connectivity}
\title{Hydrospatial analysis of hydrologic connectivity}
\usage{
hsa_connectivity(rs_ti0, fdf, wy, cres, aconv, connpoly, outdir)
}
\arguments{
\item{rs_ti0}{Raster stack or brick with inundated cells = 1}

\item{wy}{Water year to add to filenames}

\item{cres}{Resolution of cell (in units squared)}

\item{aconv}{Conversion factor for calculating area}

\item{connpoly}{Shapefile covering areas that count as connected to the river
low-flow channel}

\item{outdir}{Directory for writing rasters to file}

\item{flws}{Flows data frame for water year in format of 'utils_hsaflws'
function}
}
\value{
Flows data frame with metrics filled in. Writes rasters with
  groupings of inundation to file in the outdir.
}
\description{
Accepts binary daily rasters with inundated (1) and dry (0) cells (from
\code{hsa_extent}). For each raster, patches of inundated areas are
identified that intersect with the shapefile ('connpoly') that represents
area that should be considered as connected to the river.
Connected/disconnected inundated area rasters are written to file and
connected/disconnected area is calculated.
}
\details{
Input rs_ti0 rasters (inundated cells prior to threshold
  application) from the 'rsi' directory (could also use rs_i0, which
  represented inundated areas that meet depth and velocity threshold
  criteria). Rasters are written to file in directories 'rsc' and 'rsdc'
  within 'outdir'.
}
