% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_durationwgt.R
\name{hsa_durationwgt}
\alias{hsa_durationwgt}
\title{Hydrospatial application of inundation duration weight}
\usage{
hsa_durationwgt(rs_dayinun, durshrt, durlng, durshrt_wgt, durlng_wgt,
  durtoolng_wgt, wy, ncor, outdir)
}
\arguments{
\item{rs_dayinun}{Raster stack or brick of inundation day as from
'hsa_duration'}

\item{durshrt}{Threshold number of days for short duration inundation}

\item{durlng}{Threshold number of days for long duration inundation}

\item{durshrt_wgt}{Weight for short duration inundation}

\item{durlng_wgt}{Weight for long duration inundation}

\item{durtoolng_wgt}{Weight for too-long duration inundation}

\item{wy}{Water year to add to filenames}

\item{ncor}{Number of cores for parallel processing}

\item{outdir}{Directory for writing rasters to file}
}
\value{
Writes the duration weight rasters to file in the outdir.
}
\description{
Accepts daily rasters of inundation duration (from \code{hsa_duration}) and
applies duration weights.
}
\details{
Input rasters from 'rsdayinun0' directory. Rasters are written to file in
  a 'rshab' directory within 'outdir'.
}
