% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rasterreord.R
\name{utils_rasterreord}
\alias{utils_rasterreord}
\title{Reorder raster files}
\usage{
utils_rasterreord(fls)
}
\arguments{
\item{fls}{Vector of filenames with numbers to order after the last "_"}
}
\value{
Reordered filenames
}
\description{
Reorders raster files based on parsing vector of filenames.
}
